+++ app/models/protonym.rb
@@ -17,6 +17,10 @@ class Protonym < TaxonName
   include Protonym::SoftValidationExtensions::Instance
   include Protonym::Becomes
 
+  # @return [Boolean]
+  #   memoize `#is_avaiable?`
+  attr_reader :is_available
+
   alias_method :original_combination_source, :source
 
   FAMILY_GROUP_ENDINGS = %w{ini ina inae idae oidae odd ad oidea}.freeze
@@ -223,8 +227,13 @@ class Protonym < TaxonName
     return list
   end
 
-  def is_available?
-    !has_misspelling_relationship? && !name_is_misapplied? && !classification_invalid_or_unavailable?
+  def is_available?(refresh = false)
+    if !@refresh
+      @is_available ||= !has_misspelling_relationship? && !name_is_misapplied? && !classification_invalid_or_unavailable?
+    else
+      @is_available = !has_misspelling_relationship? && !name_is_misapplied? && !classification_invalid_or_unavailable?
+    end
+    @is_available
   end
 
   # @return [Protonym]
@@ -364,10 +373,10 @@ class Protonym < TaxonName
     else
       return nil
 
-#      RANKS.index(rank_string) <= RANKS.index(parent.rank_string)
-#      Ranks.lookup(:iczn, 'species')
-# .valid_name_ending
-#FAMILY_AND_ABOVE_RANK_NAMES
+  # RANKS.index(rank_string) <= RANKS.index(parent.rank_string)
+  #     Ranks.lookup(:iczn, 'species')
+  # .valid_name_ending
+  #FAMILY_AND_ABOVE_RANK_NAMES
     end
     r.constantize
   end
@@ -936,16 +945,15 @@ class Protonym < TaxonName
     is_cached = false if cached_author_year != get_author_and_year
 
     if is_cached && cached_html != get_full_name_html ||
-      cached_misspelling != get_cached_misspelling ||
-      cached_original_combination != get_original_combination ||
-      cached_original_combination_html != get_original_combination_html ||
-      cached_primary_homonym != get_genus_species(:original, :self) ||
-      cached_primary_homonym_alternative_spelling != get_genus_species(:original, :alternative) ||
-      rank_string =~ /Species/ &&
+        cached_misspelling != get_cached_misspelling ||
+        cached_original_combination != get_original_combination ||
+        cached_original_combination_html != get_original_combination_html ||
+        cached_primary_homonym != get_genus_species(:original, :self) ||
+        cached_primary_homonym_alternative_spelling != get_genus_species(:original, :alternative) ||
+        rank_string =~ /Species/ &&
         (cached_secondary_homonym != get_genus_species(:current, :self) ||
-          cached_secondary_homonym_alternative_spelling != get_genus_species(:current,
-                                                                             :alternative))
-      is_cached = false
+         cached_secondary_homonym_alternative_spelling != get_genus_species(:current, :alternative))
+    is_cached = false
     end
 
     soft_validations.add(