+++ lib/soft_validation.rb
@@ -1,36 +1,55 @@
 require 'soft_validation/soft_validation'
 require 'soft_validation/soft_validations'
 require 'soft_validation/soft_validation_method'
+require_relative 'utilities/params'
+require "active_support/all"
+
+# TODO: REMOVE
+require 'byebug'
+require 'amazing_print'
 
 # Vaguely inspired by concepts from by svn://rubyforge.org/var/svn/softvalidations, but not as elegant.
 #
 # Soft validations are a means to tie warnings or suggestions to instances of data.
 # Soft validations do not prevent an instance from being saved.  They are not intended
-# to be bound to AR callbacks, but this may be possible.  They may be used to
+# to be bound to AR callbacks, but this may be possible ultimately. They may be used to
 # alert the user to data issues that need to be addressed, or alert the programmer
 # who is batch parsing data as to the quality of the incoming data, etc..
 #
-# For example, soft validations could be shown on #show controller methods.
+# There are 2 stages to defining a soft validation. First index and provide an option general description
+# of the soft validation itself using the `soft_validate` macro in the class.  Second, add the method (logic)
+# that is called, and a set of message that the user will see when the logic passes or fails.  To state in another way:
+#
+# * The name and description (intent) of the soft validation is optionally provided with the macro setting the soft validation (`Klass.soft_validate()`.
+# * The human messages ('there is a problem here!', 'the problem is fixed', 'we tried to fix, but failed!') are defined with the method logic itself.  This is intentionally done to
+# keep the intent of the logic close to the consequences of the logic.
+#
+# Devloper tips:
+# 
+# - Protonym.soft_validation( ) <- all technical metadata and a gross description (the intent), optionally, goes here
+# - @protonym.sv_xyz( ) <- all human guidance (warning, outcomes) goes here, including the attribute to point to in the UI
+# - *fix* method names should not be exposed to the UI
+#
 #
 # Usage:
 #
 #   class Foo < ApplicationRecord
 #     include SoftValidation
-#     soft_validate(:a_soft_validation_method )
+#     soft_validate(:a_soft_validation_method, fix: :cook_cheezburgers)
 #
 #     # Validations can be assigned to a set (only one), and validations in a set
 #     # can be called individually.
 #     soft_validate(:other_soft_validation_method, set: :some_set)
 #     soft_validate(:yet_another_method, set: :some_other_set )
-#     soft_validate(:a_third_method, resolution: [:route_name, route_name2]) # resolution is a pointer to route/interface that can resolve the problem !! NOT TESTED
-
-#     soft_validate(:a_fourth_example, has_fix: false) # there are no fix methods assigned in :a_fourth_example 
+#     soft_validate(:described_method, name: 'the validation for X', description: 'this validation does Z')
+#     soft_validate(:a_third_method, resolution: [:route_name, route_name2]) # TODO: resolution is a pointer to route/interface that can resolve the problem !! NOT TESTED
 #
-#     $hungry = true
+#     soft_validate(:a_fourth_example, fix: :fix_method) # the detected issue can be fully resolved by calling this instance method
+#
+#     $hungry = true # demo only, don't use $globals
 #
 #     def a_soft_validation_method
 #       soft_validations.add(:base, 'hungry!',                          # :base or a model attribute (column)
-#         fix: :cook_cheezburgers,
 #         success_message: 'no longer hungry, cooked a cheezeburger',
 #         failure_message: 'oh no, cat ate your cheezeburger'
 #       ) if $hungry
@@ -79,27 +98,47 @@ require 'soft_validation/soft_validation_method'
 #
 #   f.clear_soft_validations
 #
-#   f.soft_validate(:some_other_set)          # only run this set of validations
+#   f.soft_validate(only_set: [:default])               # only run this set, that is the set of soft validations not otherwise assigned to set
+#   f.soft_validate(only_set: [:some_other_set])        # only run this set of validations
+#   f.soft_validate(except_set: [:some_other_set])      # run all except this soft validation
+#   f.soft_validate(only_method: :some_method)          # run only this method name (all other params ignored)
+#   f.soft_validate(except_method: [:some_method])      # run result except these soft validation methods
+#   f.soft_validate(fixable: true)                      # run all, but only fixable soft validations
+#   f.soft_validate(fixable: false)                     # run all, but NOT fixable soft validations
+#   f.soft_validate(flagged: true)                      # run all, *including* methods flagged by developers as "a-typical", there is no flagged: false, as it is default)
 #
 module SoftValidation
 
   class SoftValidationError < StandardError; end
 
+  # An index of the soft validators in superclasses
   ANCESTORS_WITH_SOFT_VALIDATIONS =
     Hash.new do |h, klass|
       h[klass.name] = (klass.ancestors.select {|a| a.respond_to?(:soft_validates?) && a.soft_validates?} - [klass]) # a < ApplicationRecord && would be faster but requires AR in spec
     end
 
-  extend ActiveSupport::Concern
+  extend ::ActiveSupport::Concern
 
   included do
     attr_accessor :soft_validation_result
 
+    # @return [Hash]
+    #   An index of soft validation methods by ClassName
+    #   ' { ClassName' => { method_name: @method_instance, ... }, ... }
     class_attribute :soft_validation_methods, instance_writer: false  # http://api.rubyonrails.org/classes/Class.html
+
     self.soft_validation_methods = {self.name => {}}
 
+    # @return [Hash]
+    #   An index of soft validation methods by ClassName by set
+    #   ' { ClassName' => { set: [ :method_name, ], ...}
     class_attribute :soft_validation_sets, instance_writer: false
-    self.soft_validation_sets = { self.name =>  {all:  [] } }
+    self.soft_validation_sets = { self.name =>  { default: []} }
+
+    # @return [Array]
+    #   An index of softvalidation methods
+    class_attribute :soft_validation_method_names
+    self.soft_validation_method_names = []
   end
 
   module ClassMethods
@@ -116,10 +155,12 @@ module SoftValidation
       true
     end
 
-    # @param [Symbol] method
+    # @param method [Symbol]
+    #   the name of the method with the soft validation logic, in TW like `sv_foo`
     # @param [Hash] options
     # @return [SoftValidationMethod]
     def add_method(method, options)
+      self.soft_validation_method_names.push method
       n = self.name
       self.soft_validation_methods[n] ||= {}
       self.soft_validation_methods[n][method] = SoftValidationMethod.new(options)
@@ -133,12 +174,12 @@ module SoftValidation
 
       self.soft_validation_sets[n] ||= {}
 
-      self.soft_validation_sets[n][:all] ||= []
-      self.soft_validation_sets[n][:all] << method
-
       if set
         self.soft_validation_sets[n][set] ||= []
         self.soft_validation_sets[n][set] << method
+      else
+        self.soft_validation_sets[n][:default] ||= []
+        self.soft_validation_sets[n][:default] << method
       end
     end
 
@@ -151,41 +192,9 @@ module SoftValidation
     # @return [Boolean]
     #    true if at least on soft_validate() exists in *this* class
     def has_self_soft_validations?
-      self.soft_validation_sets[self.name] && self.soft_validation_sets[self.name][:all].count > 0
-    end
-
-    # an internal accessor for self.soft_validation_methods
-    # @param [Symbol] set the set to return
-    # @param [Boolean] ancestors whether to also return the ancestors validation methods
-    # @return [Array] of Symbol
-    #   the names of the soft validation methods
-    def soft_validators(set: :all, include_ancestors: true, fixable_only: false)
-      methods = []
-     
-      klass_validators = []
-
-      
-      
-      if has_self_soft_validations?
-        a = self.soft_validation_sets[self.name][set]
-      
-        if fixable_only
-          a.each do |m|
-            klass_validators.unshift(m) if self.soft_validation_methods[self.name][m].fixable? 
-          end
-        else
-          klass_validators = a if a # self.soft_validation_sets[self.name][set] if has_self_soft_validations?
-        end
-      end
-
-      methods += klass_validators # if !klass_validators.nil?
-
-      if include_ancestors
-        ancestor_klasses_with_validation.each do |klass|
-          methods += klass.soft_validators(set: set, include_ancestors: false, fixable_only: fixable_only)
-        end
-      end
-      methods
+      self.soft_validation_method_names.any?
+      #  self.soft_validation_methods[self.name].any? # donesn't include superclass !?
+      #  self.soft_validation_sets[self.name] && self.soft_validation_sets[self.name].keys.count > 0 # [:all].count > 0 any set indicates at least one
     end
 
     # @return [Hash]
@@ -202,14 +211,92 @@ module SoftValidation
       ANCESTORS_WITH_SOFT_VALIDATIONS[self]
     end
 
+    # @param only_sets [Array]
+    #   names (symbols) of sets to run
+    #
+    # @param except_sets [Array]
+    #   names (symbols]
+    #
+    # @param only_methods [Array]
+    #   Names (symbols) of sets to run. _If provided all other params are ignored._
+    #
+    # @param except_methods [Array]
+    #   names (symbols]
+    #
+    # @param include_superclass [Boolean]
+    #   include validations on superclasses, default is `true`
+    #
+    # @param include_flagged [Boolean]
+    #   some soft validations have more consequences, these are flagged, default `false`
+    #
+    # @param fixable [Boolean]
+    #   run soft validations only on fixable records, default is `false`
+    #
+    # @return [Array] of Symbol
+    #   the names of the soft validation methods
+    #
+    #  An internal accessor for self.soft_validation_methods.  If nothing is provided all possible specs, excluding those flagged are returned.
+    def soft_validators(only_sets: [], except_sets: [], only_methods: [], except_methods: [], include_flagged: false, fixable: nil, include_superclass: true)
+      only_methods = ::Utilities::Params.arrayify(only_methods)
+      return only_methods if !only_methods.empty?
+
+      except_methods = ::Utilities::Params.arrayify(except_methods)
+
+      # Get sets
+      sets = get_sets(
+        ::Utilities::Params.arrayify(only_sets),
+        ::Utilities::Params.arrayify(except_sets)
+      )
+
+      methods = []
+      klass_validators = []
+
+      # Return "Local" (this class only) validators
+      if has_self_soft_validations?
+
+        a = []
+        if sets.empty?
+          a = self.soft_validation_method_names
+        else
+          sets.each do |s|
+            a += self.soft_validation_sets[self.name][s]
+          end
+        end
+
+        a.delete_if{|n| !self.soft_validation_methods[self.name][n].send(:matches?, fixable, include_flagged) }
+        methods += a
+      end
+
+      # Add the rest of the validators, from Superclasses
+      if include_superclass
+        ancestor_klasses_with_validation.each do |klass|
+          methods += klass.soft_validators(include_superclass: false, only_sets: only_sets, except_sets: except_sets, except_methods: except_methods, include_flagged: include_flagged, fixable: fixable)
+        end
+      end
+
+      # Get rid of explicitly excluded
+      methods.delete_if{|m| except_methods.include?(m) }
+
+      methods
+    end
+
+    private
+
     def reset_soft_validation!
       self.soft_validation_methods = {self.name => {}}
-      self.soft_validation_sets = { self.name =>  {all:  [] } }
+      self.soft_validation_sets = { self.name => { default: []}}
+      self.soft_validation_method_names = []
+    end
+
+    def get_sets(only_sets = [], except_sets = [])
+      all_sets = soft_validation_sets[name].keys
+      a = (all_sets - except_sets)
+      only_sets.empty? ? a : a & only_sets
     end
 
   end
 
-  # instance methods
+  # Instance methods
 
   # @return [SoftValidations]
   def soft_validations
@@ -221,52 +308,49 @@ module SoftValidation
     @soft_validation_result = nil
   end
 
-  # @param [Symbol] set the set of soft validations to run
-  # @param [Boolean] ancestors whether to also validate ancestors soft validations
-  # @return [Boolean] always true
-  def soft_validate(set = :all, include_ancestors = true, only_fixable = false)
+  # Run a set of soft validations.
+  # * by default all validations except those with `flagged: true` are run
+  # * when only|except_methods are set then these further restrict the scope of tests run
+  # * except_methods will exclude methods from *any* result (i.e. sets are allowed)
+  #
+  # @param (see SoftValidation#soft_validators)
+  #
+  # @return [true]
+  def soft_validate(**options)
     clear_soft_validations
     soft_validations
-    sets = case set.class.name
-           when 'Array'
-             set
-           when 'Symbol'
-             [set]
-           when 'String'
-             [set.to_sym]
-           end
 
-    sets.each do |s|
-      self.class.soft_validators(set: s, include_ancestors: include_ancestors, fixable_only: only_fixable).each do |m|
-        self.send(m)
-      end
+    soft_validators(**options).each do |sv_method|
+      self.send(sv_method)
     end
+
     soft_validations.validated = true
     true
   end
 
-  # @param [Symbol] scope
-  # @return [Boolean]
-  def fix_soft_validations(scope = :automatic)
+  # @see Class.soft_validators
+  def soft_validators(**options)
+    self.class.soft_validators(**options)
+  end
+
+  # The validation set to fix is set prior to running the fix, at the first step.
+  # It can be refined/restricted there as needed, letting specific contexts (e.g. 
+  # access in controller) defined the scope.
+  def fix_soft_validations
     return false if !soft_validated?
-    raise 'invalid scope passed to fix_soft_validations' if ![:automatic, :requested].include?(scope)
 
     soft_validations.soft_validations.each do |v|
       if v.fix
-        if v.fix_trigger == scope
           if self.send(v.fix)
             v.fixed = :fixed
           else
             v.fixed = :fix_error
           end
-        else
-          v.fixed = :fix_not_triggered
-        end
       else
         v.fixed = :no_fix_available
       end
     end
-    soft_validations.fixes_run = scope
+    soft_validations.fixes_run = true
     true
   end