+++ spec/lib/soft_validation/soft_validations_spec.rb
@@ -22,16 +22,11 @@ describe 'SoftValidations', group: :soft_validation do
     expect(soft_validations.soft_validations.count).to eq(1)
   end
 
-  specify 'add(:attribute, "message", fix: :method, fix_trigger: :automatic)' do
-    expect(soft_validations.add(:base, 'no cheezburgahz!', fix: 'cook_a_burgah', fix_trigger: :automatic)).to be_truthy
-    expect(soft_validations.soft_validations.count).to eq(1)
-  end
-
   specify 'add with success/fail message without fix returns false' do
     expect(soft_validations.add(:base,'no cheezburgahz!', success_message: 'cook_a_burgah')).to be_falsey
   end
 
-  specify 'add(:attribute, "message", fix: :method, success_message: "win",  failure_message: "fail")' do
+  specify 'add(:attribute, "message", success_message: "win",  failure_message: "fail")' do
     expect(soft_validations.add(:base, 'no cheezburgahz!', fix: 'cook_a_burgah', success_message: 'haz cheezburger', failure_message: 'no cheezburger')).to be_truthy
   end
 
@@ -56,6 +51,27 @@ describe 'SoftValidations', group: :soft_validation do
   specify '#resolution_for' do
     expect(soft_validations).to respond_to(:resolution_for)
   end
+
+  specify '#validated?' do
+    expect(soft_validations.validated?).to eq(false)
+  end 
+
+  specify '#fixes_run?' do
+    expect(soft_validations.fixes_run?).to eq(false)
+  end 
+
+  specify '#complete?' do
+    expect(soft_validations.complete?).to eq(false)
+  end 
+
+  specify '#fix_messages' do
+    expect(soft_validations.fix_messages).to eq({})
+  end 
+
+  specify '#size' do
+    soft_validations.add(:base, 'no cheezburgahz!')
+    expect(soft_validations.size).to eq(1)
+  end 
 
 end