+++ spec/lib/soft_validation_spec.rb
@@ -1,10 +1,9 @@
-require 'rails_helper'
+# require 'rails_helper'
 require 'lib/soft_validation_helpers'
 
 describe 'SoftValidation', group: :soft_validation do
 
   context 'extended class methods' do
-
     specify '.has_self_soft_validations?' do
       expect(Softy.has_self_soft_validations?).to be_falsey
     end
@@ -27,58 +26,62 @@ describe 'SoftValidation', group: :soft_validation do
 
     context 'adding soft validations' do
       before(:each) {
-        Softy.reset_soft_validation!
+        Softy.send(:reset_soft_validation!)
       }
 
-      specify 'basic use: soft_validate(:method)' do
-        expect(Softy.soft_validate(:haz_cheezburgers?)).to be_truthy
+      specify 'basic use: soft_validate(:method) 1' do
+        Softy.soft_validate(:haz_cheezburgers?)
         expect(Softy.soft_validation_methods['Softy'].keys).to contain_exactly(:haz_cheezburgers?)
+      end
+
+      specify '#soft_validation_method_names' do
+        Softy.soft_validate(:haz_cheezburgers?)
+        expect(Softy.soft_validation_method_names).to contain_exactly(:haz_cheezburgers?)
       end
 
       specify 'assigment to a named set of validations: soft_validate(:method, set: :set_name)' do
         expect(Softy.soft_validate(:needs_moar_cheez?, set: :cheezy)).to be_truthy
-        expect(Softy.soft_validation_sets['Softy'][:all]).to contain_exactly(:needs_moar_cheez?)
+        #  expect(Softy.soft_validation_sets['Softy'][:all]).to contain_exactly(:needs_moar_cheez?)
         expect(Softy.soft_validation_sets['Softy'][:cheezy]).to contain_exactly(:needs_moar_cheez?)
       end
 
       specify 'assigment with name, description, resolution' do
         expect(Softy.soft_validate(:needs_moar_cheez?, name: 'Check for cheeze', description: 'Open bun, look for cheeze.', resolution: [:root_path])).to be_truthy
-        expect(Softy.soft_validation_methods['Softy'][:needs_moar_cheez?].name).to eq('Check for cheeze') 
-        expect(Softy.soft_validation_methods['Softy'][:needs_moar_cheez?].description).to eq('Open bun, look for cheeze.') 
-        expect(Softy.soft_validation_methods['Softy'][:needs_moar_cheez?].resolution).to contain_exactly(:root_path) 
+        expect(Softy.soft_validation_methods['Softy'][:needs_moar_cheez?].name).to eq('Check for cheeze')
+        expect(Softy.soft_validation_methods['Softy'][:needs_moar_cheez?].description).to eq('Open bun, look for cheeze.')
+        expect(Softy.soft_validation_methods['Softy'][:needs_moar_cheez?].resolution).to contain_exactly(:root_path)
       end
 
       specify 'assignment with name, fixable' do
-        expect(Softy.soft_validate(:needs_moar_cheez?, name: 'Check for cheeze', has_fix: false)).to be_truthy
-        expect(Softy.soft_validation_methods['Softy'][:needs_moar_cheez?].has_fix).to eq(false) 
-        expect(Softy.soft_validation_methods['Softy'][:needs_moar_cheez?].fixable?).to eq(false) 
+        expect(Softy.soft_validate(:needs_moar_cheez?, name: 'Check for cheeze')).to be_truthy
+        expect(Softy.soft_validation_methods['Softy'][:needs_moar_cheez?].fixable?).to eq(false)
       end
     end
 
-    context 'fixable methods' do
-      before(:each) {
-        Softy.reset_soft_validation!
-        expect(Softy.soft_validate(:needs_moar_cheez?, name: 'Check for cheeze')).to be_truthy
-        expect(Softy.soft_validate(:needs_less_cheez?, name: 'Too much cheeze', has_fix: false)).to be_truthy
-      }
+    context '#fix, i.e. fixable soft validations' do
+      before(:each) do
+        Softy.send(:reset_soft_validation!)
+        Softy.soft_validate(:needs_moar_cheez?, name: 'Check for cheeze')
+        Softy.soft_validate(:needs_less_cheez?, name: 'Too much cheeze', fix: :remove_cheeze)
+      end
 
       specify '.soft_validators default are all' do
-        expect(Softy.soft_validators).to contain_exactly(:needs_moar_cheez?, :needs_less_cheez?) 
+        expect(Softy.soft_validators).to contain_exactly(:needs_moar_cheez?, :needs_less_cheez?)
       end
 
       specify '.soft_validators, fixable_only' do
-        expect(Softy.soft_validators(fixable_only: true)).to contain_exactly(:needs_moar_cheez?) 
+        expect(Softy.soft_validators(fixable: true)).to contain_exactly(:needs_less_cheez?)
       end
 
       specify '.soft_validators false is all' do
-        expect(Softy.soft_validators(fixable_only: false)).to contain_exactly(:needs_moar_cheez?, :needs_less_cheez?) 
+        expect(Softy.soft_validators(fixable: false)).to contain_exactly(:needs_moar_cheez?)
       end
     end
 
     context 'soft_validations between classes' do
       before(:each) {
-        Softy.reset_soft_validation!
-        OtherSofty.reset_soft_validation!
+        Softy.send(:reset_soft_validation!)
+        OtherSofty.send(:reset_soft_validation!)
       }
 
       specify 'methods assigned to parent are available to child' do
@@ -88,7 +91,7 @@ describe 'SoftValidation', group: :soft_validation do
 
       specify 'methods assigned to child are not merged with parents' do
         OtherSofty.soft_validate(:haz_cheezburgers?)
-        expect(Softy.soft_validation_sets['Softy'][:all]).to contain_exactly()
+        expect(Softy.soft_validation_sets['Softy'][:default]).to contain_exactly()
       end
 
       context 'accessed via soft_validators' do
@@ -136,14 +139,22 @@ describe 'SoftValidation', group: :soft_validation do
     end
   end
 
+  specify '.reset_soft_validation' do
+    Softy.soft_validate(:just_bun?)
+    Softy.send(:reset_soft_validation!)
+    expect(Softy.soft_validation_method_names).to eq([])
+    expect(Softy.soft_validation_sets).to eq( 'Softy' => {default: []})
+    expect(Softy.soft_validation_methods).to eq( 'Softy' =>  {})
+  end
+
   context 'example usage' do
-    before { Softy.reset_soft_validation! }
+    before { Softy.send(:reset_soft_validation!) }
 
     context 'with a validation that has resolution' do
-      before do 
+      before do
         Softy.soft_validate(:just_bun?, resolution: [:root_path])
         Softy.soft_validate(:needs_moar_cheez?)
-      end 
+      end
 
       let(:softy) {Softy.new}
 
@@ -173,11 +184,11 @@ describe 'SoftValidation', group: :soft_validation do
     end
 
     context 'with a couple of validations' do
-      before do 
-        # Stub the validation methods 
+      before do
+        # Stub the validation methods
         Softy.soft_validate(:needs_moar_cheez?, set: :cheezy)
         Softy.soft_validate(:haz_cheezburgers?)
-      end 
+      end
 
       let(:softy) {Softy.new}
 
@@ -206,32 +217,41 @@ describe 'SoftValidation', group: :soft_validation do
           expect(softy.soft_validations.class).to eq(SoftValidation::SoftValidations)
         end
 
-        specify '#fixes_run?' do
-          expect(softy.soft_validations.fixes_run?).to be_falsey
+        specify '#fixes_run? 1' do
+          expect(softy.soft_validations.fixes_run?).to eq(false)
         end
 
-        specify 'softy.fix_soft_validations(:some_bad_value)' do
-          expect{softy.fix_soft_validations(:some_bad_value)}.to raise_error(RuntimeError, /invalid scope/)
-        end
-
-        specify 'softy.fix_soft_validations' do
+        specify '#on' do
           expect(softy.soft_validations.on(:mohr).size).to eq(1)
+        end
+
+        specify '#fix_soft_validations' do
           expect(softy.fix_soft_validations).to be_truthy
-          expect(softy.soft_validations.fixes_run?).to eq(:automatic)
-          expect(softy.soft_validations.fix_messages).to eq(base: ['no longer hungry, cooked a cheezeburger'], mohr: ['fix not run, no automatic fix available'])
-
-          # TODO: Move out  
-          expect(softy.soft_validations.on(:mohr).size).to eq(1)
-          expect(softy.soft_validations.messages).to eq(['hungry (for cheez)!', 'hungry!'] ) 
-          expect(softy.soft_validations.messages_on(:mohr)).to eq(['hungry (for cheez)!'] ) 
         end
 
-        specify 'softy.fix_soft_validations(:requested)' do
-          expect(softy.soft_validations.on(:mohr).size).to eq(1)
-          expect(softy.fix_soft_validations(:requested)).to be_truthy
-          expect(softy.soft_validations.fixes_run?).to eq(:requested) 
-          expect(softy.soft_validations.fix_messages).to eq(mohr: ['fix not run, no automatic fix available'], base: ['fix available, but not triggered'])
+        specify '#fixes_run? 2' do
+          softy.fix_soft_validations
+          expect(softy.soft_validations.fixes_run?).to eq(true)
         end
+
+        specify '#fix_messages' do
+          softy.fix_soft_validations
+          expect(softy.soft_validations.fix_messages).to eq(base: ['no longer hungry, cooked a cheezeburger'], mohr: ['no fix available'])
+        end
+
+        specify '#messages' do
+          softy.fix_soft_validations
+          expect(softy.soft_validations.messages).to contain_exactly('hungry (for cheez)!', 'hungry!' )
+        end
+
+        specify '#messages_on' do
+          softy.fix_soft_validations
+          expect(softy.soft_validations.messages_on(:mohr)).to eq(['hungry (for cheez)!'] )
+        end
+
+        #  specify 'softy.fix_soft_validations(:requested) 4' do
+        #    expect(softy.soft_validations.fix_messages).to eq(mohr: ['no fix available'], base: ['fix available, but not triggered']) # flagged
+        #  end
 
       end
     end